Learning about HR Analytics metric - Attrition
I've just created a dynamic Tableau dashboard that serves as a fantastic playground for practicing my Tableau skills and delving into the fascinating world of HR Analytics.
Through this project I was able to learn about one of the key HR Metrics - Attrition Rate. It is a metric used in HR analytics to measure the rate at which employees leave an organization, whether voluntarily or involuntarily, over a specific period. It is usually expressed as a percentage and is used as one of the main figures HR tracks to understand how things are going at a company.
Diving further into the HR analytics, I discovered that SAP SuccessFactors is one of the top rated HCM suites for large organizations (1000+ employees) according to Gartner.
Through this project, I've honed my Tableau expertise, experimenting with various charts, graphs, and interactive elements to present data in meaningful and impactful ways. The process has been a fantastic learning experience, allowing me to blend technical skills with an understanding of HR practices.



You can check out this dashboard on my Tableau Public Profile.
https://shorturl.at/pZ578